Updated the files to download the data

+++ SOMOSPIE/data_readers/satellite/fetch_ecoregions.sh
@@ -7,30 +7,40 @@
 # Fetching Eco-Region level 1
 #
 ECO=NA_Terrestrial_Ecoregions_Level_I_Shapefile
-wget http://www.cec.org/wp-content/uploads/wpallimport/files/Atlas/Files/Terrestrial_Ecoregions_L1/NA_Terrestrial_Ecoregions_Level_I_Shapefile.zip
-unzip -o NA_Terrestrial_Ecoregions_Level_I_Shapefile.zip
-rm -r NA_Terrestrial_Ecoregions_Level_I_Shapefile.zip
-echo $(date) > $ECO/downloaded.txt
-find $ECO/ -name "NA_Terrestrial_Ecoregions_v2_level1.shp" -exec echo {} > path_CEC.txt \;
+### Check for dir, if not found create it using the mkdir ##
+
+if [ ! -d "ECO" ]; then
+
+    wget http://www.cec.org/wp-content/uploads/wpallimport/files/Atlas/Files/Terrestrial_Ecoregions_L1/NA_Terrestrial_Ecoregions_Level_I_Shapefile.zip
+    unzip -o NA_Terrestrial_Ecoregions_Level_I_Shapefile.zip
+    rm -r NA_Terrestrial_Ecoregions_Level_I_Shapefile.zip
+    echo $(date) > $ECO/downloaded.txt
+    find $ECO/ -name "NA_Terrestrial_Ecoregions_v2_level1.shp" -exec echo {} > path_CEC.txt \;
+fi
 
 # Fetching Eco-Region level 2
 #
 ECO=NA_Terrestrial_Ecoregions_Level_II_Shapefile
-wget http://www.cec.org/wp-content/uploads/wpallimport/files/Atlas/Files/Terrestrial_Ecoregions_L2/NA_Terrestrial_Ecoregions_Level_II_Shapefile.zip
-unzip -o NA_Terrestrial_Ecoregions_Level_II_Shapefile.zip
-rm -r NA_Terrestrial_Ecoregions_Level_II_Shapefile.zip
-echo $(date) > $ECO/downloaded.txt
-find $ECO/ -name "NA_Terrestrial_Ecoregions_v2_level2.shp" -exec echo {} > path_CEC.txt \;
+### Check for dir, if not found create it using the mkdir ##
+if [ ! -d "ECO" ]; then
+
+    wget http://www.cec.org/wp-content/uploads/wpallimport/files/Atlas/Files/Terrestrial_Ecoregions_L2/NA_Terrestrial_Ecoregions_Level_II_Shapefile.zip
+    unzip -o NA_Terrestrial_Ecoregions_Level_II_Shapefile.zip
+    rm -r NA_Terrestrial_Ecoregions_Level_II_Shapefile.zip
+    echo $(date) > $ECO/downloaded.txt
+    find $ECO/ -name "NA_Terrestrial_Ecoregions_v2_level2.shp" -exec echo {} > path_CEC.txt \;
+fi
 
 # Fetching Eco-Region level 3
 #
 ECO=NA_Terrestrial_Ecoregions_v2_Level_III_Shapefile
-wget http://www.cec.org/wp-content/uploads/wpallimport/files/Atlas/Files/Terrestrial_Ecoregions_L3/NA_Terrestrial_Ecoregions_v2_Level_III_Shapefile.zip
-unzip -o NA_Terrestrial_Ecoregions_v2_Level_III_Shapefile.zip
-rm -r NA_Terrestrial_Ecoregions_v2_Level_III_Shapefile.zip
-echo $(date) > $ECO/downloaded.txt
-find $ECO/ -name "NA_Terrestrial_Ecoregions_v2_level3.shp" -exec echo {} > path_CEC.txt \;
-
+if [ ! -d "ECO" ]; then
+    wget http://www.cec.org/wp-content/uploads/wpallimport/files/Atlas/Files/Terrestrial_Ecoregions_L3/NA_Terrestrial_Ecoregions_v2_Level_III_Shapefile.zip
+    unzip -o NA_Terrestrial_Ecoregions_v2_Level_III_Shapefile.zip
+    rm -r NA_Terrestrial_Ecoregions_v2_Level_III_Shapefile.zip
+    echo $(date) > $ECO/downloaded.txt
+    find $ECO/ -name "NA_Terrestrial_Ecoregions_v2_level3.shp" -exec echo {} > path_CEC.txt \;
+fi
 # The following two lines were used once to update .gitignore, which was updated in the repo,
 #  so these two lines were commented out and need not be run again.
 #echo $ECO >> .gitignore
@@ -38,7 +48,10 @@ find $ECO/ -name "NA_Terrestrial_Ecoregions_v2_level3.shp" -exec echo {} > path_
 
 NEON=NEONDomains_0
 wget https://www.neonscience.org/sites/default/files/$NEON.zip
+if [ ! -d "NEON" ]; then
 mkdir $NEON
+fi
+
 cd $NEON
 unzip -o ../$NEON.zip
 cd ../
